Set minimal permissions for pipeline
If a GitHub Actions job or workflow has no explicit permissions set, then the repository permissions are used. Repositories created under organizations inherit the organization permissions. Often these permissions do not adhere to the principle of least privilege and can be reduced to read-only. To fix the problem, we explicitly set the permissions block at the workflow level to the minimum required. This will apply to all jobs unless they override it with their own permissions block.