Update frontend dependencies
Bumping react-router from 7.7.0 to 7.8.2 seems to have caused a change in action redirect behavior, with the Remix Single Fetch strategy seemingly kicking in. While previously, the pre-check email template opening was caused by a 302 Found response, the actual network traffic now shows a 202 Accepted response, with the mailto: link opened client-side. This made it hard to intercept the `ergebnis.data` route in Playwright.

As a workaround, `page.waitForURL` is used in end-to-end tests to replace the route interception. A prefix is introduced for end-to-end tests, to prevent the local mailto: handler from opening when running E2E tests.