Replace fieldsets and legends (a11y)
Using headings and long-form content in legends is considered bad practice because it creates a lot of noise. The textfield fieldsets were removed because the labels should be sufficient, and users can explore the surrounding content for more context.
The radio button fieldsets were changed to divs with roles to associate them with heading elements that stand on their own.